Allow to mock router enabled component easily
